requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.22
pymysql==1.0.2
Werkzeug==2.0.1
python-dotenv==0.19.1

